Change Double to BigDecimal

+++ AuctionBidSystem/src/main/java/com/auction/model/auction/BidTransaction.java
@@ -1,6 +1,8 @@
 package com.auction.model.auction;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
+import com.auction.model.base.Entity;
 import com.auction.model.item.Item;
 import com.auction.model.user.Bidder;
 
@@ -10,10 +12,10 @@ public class BidTransaction extends Entity {
     private static int transactionCounter = 0;
     private final Item item;
     private final Bidder bidder;
-    private final double bidAmount;
+    private final BigDecimal bidAmount;
     private final LocalDateTime bidTimestamp;
 
-    public BidTransaction(Item item, Bidder bidder, double bidAmount, LocalDateTime bidTimestamp) {
+    public BidTransaction(Item item, Bidder bidder, BigDecimal bidAmount, LocalDateTime bidTimestamp) {
         super("BID" + (++transactionCounter));
         this.item = item;
         this.bidder = bidder;
@@ -24,6 +26,6 @@ public class BidTransaction extends Entity {
     // LẤY VÀ UPDATE THUỘC TÍNH
     public Item getItem() { return item; }
     public Bidder getBidder() { return bidder; }
-    public double getBidAmount() { return bidAmount; }
+    public BigDecimal getBidAmount() { return bidAmount; }
     public LocalDateTime getBidTimestamp() { return bidTimestamp; }
 }
